Fixed survey parameters getting obfuscated by proguard
